scripts/dgccrf-rewrite/requirements.txt
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0
tqdm>=4.65
python-frontmatter>=1.1
chromadb>=0.5
sentence-transformers>=3.0
openai>=1.40
